changes in histories

+++ src/components/settings/setting.vue
@@ -1,21 +1,5 @@
-<template>
-  <div class="settings">
-    <button class="editProfile__top-btn" @click="$router.push('/profile')"><i class="fa-solid fa-arrow-left"></i></button>
-    <button class="profileUser__header-btn" @click="closeSesion()">
-      <i class="fa fa-sign-out" aria-hidden="true"></i>
-    </button>
-  </div>
-  <p>x</p>
-  <p>x</p>
-  <p>x</p>
-  <p>x</p>
-  <p>x</p>
-  <p>x</p>
-  <p>x</p>
-  <button @click="deleteAccount()">Delete Account</button>
-</template>
-
 <script setup>
+import { ref } from "vue";
 import { useRouter } from "vue-router";
 import axios from "../../api/axios";
 const router = useRouter();
@@ -24,18 +8,60 @@ const closeSesion = () => {
   router.push("/");
   window.location.reload();
 };
+let confirmationDelete = ref(false);
+let modalConfirm = ref(false);
+
 const deleteAccount = async () => {
-  const reponse = await axios.delete("/usersModule/deleteAccount")
-  console.log(reponse.data)
-  localStorage.clear() 
-  router.push("/") 
-  window.location.reload()
+  modalConfirm.value = !modalConfirm.value;
+  if(!modalConfirm.value){
+    confirmationDelete.value = !confirmationDelete.value
+  }
+  if (confirmationDelete.value) {
+    const reponse = await axios.delete("/usersModule/deleteAccount");
+    console.log(reponse.data);
+    localStorage.clear();
+    router.push("/");
+    window.location
+      .reload()
 
-    .catch((error) => {
-      push.error({});
-    });
-}; 
+      .catch((error) => {
+        push.error({});
+      });
+  }
+};
+// function yesDelete(){
+//   confirmationDelete.value = !confirmationDelete.value
+//   modalConfirm.value = !modalConfirm.value;
+// }
+function notDelete() {
+  modalConfirm.value = !modalConfirm.value;
+}
+</script>
 
-</script> 
+<template>
+  <div class="settings">
+    <div class="settings__header">
+      <p>Configuración</p>
+    </div>
+    <button class="editProfile__top-btn" @click="$router.push('/profile')"><i class="fa-solid fa-arrow-left"></i></button>
+    <button class="profileUser__header-btn" @click="closeSesion()">
+      <i class="fa fa-sign-out" aria-hidden="true"></i>
+    </button>
 
-<style></style>
+    <div v-if="modalConfirm">
+      <p>Estas segudo de eliminar tu cuenta?</p>
+      <button @click="deleteAccount(true)">Confirmar</button>
+      <button @click="notDelete()">Cancelar</button>
+    </div>
+    <p>x</p>
+    <p>x</p>
+    <p>x</p>
+    <p>x</p>
+    <p>x</p>
+    <p>x</p>
+    <p>x</p>
+    <button @click="deleteAccount()">Delete Account</button>
+  </div>
+</template>
+
+<style lang="scss"></style>
